refactor: refactor define typography constants
