+++ .storybook/preview.ts
@@ -1,8 +1,4 @@
 import { themes } from '@storybook/theming';
-import '@fontsource/roboto/300.css';
-import '@fontsource/roboto/400.css';
-import '@fontsource/roboto/500.css';
-import '@fontsource/merriweather/400.css';
 
 const themeDefaults = {
   brandTitle: 'Adaptive UI Kit 2022',
