+++ src/index.ts
@@ -1,4 +1,9 @@
+import '@fontsource/roboto/300.css';
+import '@fontsource/roboto/400.css';
+import '@fontsource/roboto/500.css';
+import '@fontsource/merriweather/400.css';
 export { Example } from './components/example';
 export { LogoIcon } from './components/LogoIcon';
 export { LogoText } from './components/LogoText';
 export { MaximizedLogo } from './components/MaximizedLogo';
+export { Typography } from './components/Typography';
