+++ src/theme/index.ts
@@ -1,12 +1,14 @@
 import { createTheme } from '@mui/material/styles';
-import * as palette from './palette';
 
-// See: https://mui.com/material-ui/customization/default-theme/
+import * as palette from './palette';
+import { THEME_TYPOGRAPHY } from './typography';
 
 export const light = createTheme({
-  palette: palette.light
+  palette: palette.light,
+  typography: THEME_TYPOGRAPHY,
 });
 
 export const dark = createTheme({
-  palette: palette.dark
+  palette: palette.dark,
+  typography: THEME_TYPOGRAPHY,
 });
